+++ CHANGELOG.md
@@ -770,6 +770,8 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#1953 Uniqueness validation now works on localized fields.
+
 * \#1936 Allow setting n levels deep embedded documents atomically without
   conflicting mods when not using nested attributes or documents themselves
   in an update call from the parent.